Added info to the Blog.sol file

+++ contracts/Blog.sol
@@ -13,20 +13,24 @@ contract Blog {
     Counters.Counter private _postIds;
 
     struct Post {
-      uint id;
-      string title;
-      string content;
-      bool published;
+        uint256 id;
+        string title;
+        string content;
+        bool published;
     }
+
     /* mappings can be seen as hash tables */
     /* here we create lookups for posts by id and posts by ipfs hash */
-    mapping(uint => Post) private idToPost;
+    /* The IPFS (InterPlanetary File System) is a protocol and peer-to-peer network for storing and sharing data in a distributed file system. */
+    /* IPFS uses content-addressing to uniquely identify each file in a global namespace connecting all computing devices. */
+
+    mapping(uint256 => Post) private idToPost;
     mapping(string => Post) private hashToPost;
 
     /* events facilitate communication between smart contractsand their user interfaces  */
     /* i.e. we can create listeners for events in the client and also use them in The Graph  */
-    event PostCreated(uint id, string title, string hash);
-    event PostUpdated(uint id, string title, string hash, bool published);
+    event PostCreated(uint256 id, string title, string hash);
+    event PostUpdated(uint256 id, string title, string hash, bool published);
 
     /* when the blog is deployed, give it a name */
     /* also set the creator as the owner of the contract */
@@ -47,14 +51,17 @@ contract Blog {
     }
 
     /* fetches an individual post by the content hash */
-    function fetchPost(string memory hash) public view returns(Post memory){
-      return hashToPost[hash];
+    function fetchPost(string memory hash) public view returns (Post memory) {
+        return hashToPost[hash];
     }
 
     /* creates a new post */
-    function createPost(string memory title, string memory hash) public onlyOwner {
+    function createPost(string memory title, string memory hash)
+        public
+        onlyOwner
+    {
         _postIds.increment();
-        uint postId = _postIds.current();
+        uint256 postId = _postIds.current();
         Post storage post = idToPost[postId];
         post.id = postId;
         post.title = title;
@@ -65,8 +72,13 @@ contract Blog {
     }
 
     /* updates an existing post */
-    function updatePost(uint postId, string memory title, string memory hash, bool published) public onlyOwner {
-        Post storage post =  idToPost[postId];
+    function updatePost(
+        uint256 postId,
+        string memory title,
+        string memory hash,
+        bool published
+    ) public onlyOwner {
+        Post storage post = idToPost[postId];
         post.title = title;
         post.published = published;
         post.content = hash;
@@ -76,12 +88,12 @@ contract Blog {
     }
 
     function fetchPosts() public view returns (Post[] memory) {
-        uint itemCount = _postIds.current();
-        uint currentIndex = 0;
+        uint256 itemCount = _postIds.current();
+        uint256 currentIndex = 0;
 
         Post[] memory posts = new Post[](itemCount);
-        for (uint i = 0; i < itemCount; i++) {
-            uint currentId = i + 1;
+        for (uint256 i = 0; i < itemCount; i++) {
+            uint256 currentId = i + 1;
             Post storage currentItem = idToPost[currentId];
             posts[currentIndex] = currentItem;
             currentIndex += 1;
@@ -92,7 +104,7 @@ contract Blog {
     /* this modifier means only the contract owner can */
     /* invoke the function */
     modifier onlyOwner() {
-      require(msg.sender == owner);
-    _;
-  }
+        require(msg.sender == owner);
+        _;
+    }
 }
